[benc/item-flipbook] Move ItemFlipbook to its own file
